+++ src/problems/Euler2.cpp
@@ -12,19 +12,19 @@ int euler<2>() {
 
   int p_num = 2;
 
-  Filter<int> fil([] (const int &x)
-    { return ! (x%2); });
+  Filter<long long> fil([] (const long long &x)
+    { return ! (x%2LL); });
 
   Fibonacci fib;
 
   auto sve = fil.wrap(fib);
 
-  Cap<int> capi([] (const int &x)
-    { return x < 4000000; });
+  Cap<long long> capi([] (const long long &x)
+    { return x < 4000000LL; });
 
   auto fin = capi.wrap(sve);
 
-  answer_report(p_num, accumulate(fin.begin(), fin.end(), 0));
+  answer_report(p_num, accumulate(fin.begin(), fin.end(), 0LL));
 
   return 0;
