+++ src/problems/Euler6.cpp
@@ -9,11 +9,11 @@ int euler<6>() {
 
   int p_num = 6;
 
-  Range r(1,101);
+  Range r(1LL,101LL);
 
-  auto f = [] (const int &x, const int &y) { return x - y*y; };
+  auto f = [] (const long long &x, const long long &y) { return x - y*y; };
 
-  int v1 = accumulate(r.begin(), r.end(),0);
+  long long v1 = accumulate(r.begin(), r.end(), 0LL);
 
   answer_report(p_num, accumulate(r.begin(),r.end(),v1 *= v1,f));
